rename logger header
