+++ stitcher/include/airmap/logger_moo.h
